requirements.txt
aiogram==3.7.0
redis==5.0.4
Pillow==10.3.0
APScheduler==3.10.4